+++ chartofaccountDIT/test/test_view_archived_programme.py
@@ -13,7 +13,7 @@ from chartofaccountDIT.views import (
     HistoricalFilteredProgrammeView,
 )
 
-from core.myutils import get_current_financial_year
+from core.generic_utils import get_current_financial_year
 from core.test.test_base import RequestFactoryBase
