+++ chartofaccountDIT/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render
 
 from core.models import FinancialYear
-from core.myutils import get_current_financial_year
+from core.generic_utils import get_current_financial_year
 from core.views import FAdminFilteredView, HistoricalFilteredView
 
 from .filters import (
