+++ core/test/test_utils.py
@@ -7,7 +7,7 @@ from django.test import (
 
 from core.models import FinancialYear
 from core.test.test_base import RequestFactoryBase
-from core.utils import log_object_change
+from core.generic_utils import log_object_change
 
 
 class ViewUtils(TestCase):
