+++ core/views.py
@@ -10,11 +10,11 @@ from django_tables2.views import SingleTableMixin
 
 from core.exportutils import EXC_TAB_NAME_LEN
 from core.models import Document
-from core.myutils import (
+from core.generic_utils import (
     get_current_financial_year,
     get_year_display,
 )
-from core.utils import today_string
+from core.generic_utils import today_string
 
 
 @login_required()
