+++ costcentre/test/test_archive.py
@@ -6,7 +6,7 @@ from django.core.management import call_command
 from django.test import TestCase
 from django.urls import reverse
 
-from core.myutils import get_current_financial_year
+from core.generic_utils import get_current_financial_year
 from core.test.test_base import RequestFactoryBase
 
 from costcentre.test.factories import (
