+++ forecast/test/test_download_edit_forecast.py
@@ -13,7 +13,7 @@ from chartofaccountDIT.test.factories import (
 )
 
 from core.models import FinancialYear
-from core.myutils import get_current_financial_year
+from core.generic_utils import get_current_financial_year
 from core.test.test_base import RequestFactoryBase
 
 from costcentre.test.factories import (
