+++ forecast/utils/export_helpers.py
@@ -11,7 +11,7 @@ from core.exportutils import (
     EXCEL_TYPE,
     EXC_TAB_NAME_LEN,
 )
-from core.utils import today_string
+from core.generic_utils import today_string
 
 from forecast.models import FinancialPeriod
 from forecast.utils.view_header_definition import (
